diag(parity): arc segmentation identical per GT per-arc dump; residual is per-segment zero-block filtering
GT arcseg patch (true nix-source context; the first attempt mislanded
in the planner — nixpkgs source layout differs from the repo tree,
ENTER probe verified 13368/13368 process_G2_G3 entries) dumps per-arc
(n, r, a): 145440 segments both sides, zero count diffs. The +9919
block delta is a dump-scope artifact (GT cache pushes 113894 vs ares
all-blocks 123813); the real residual is which identical segments get
a planner block at all.

+++ crates/ares-core/src/project_slice/gcode_emit/processor/arc_accounting.rs
@@ -45,8 +45,12 @@ pub(super) fn arc_internal_g1_lines(code: &str, command: &str, state: &MotionSta
             }
         }
         None => {
-            let i = word(code, 'I').unwrap_or(0.0);
-            let j = word(code, 'J').unwrap_or(0.0);
+            // `Vec3f rel_center` (`GCodeProcessor.cpp:4590-4596`): the I/J
+            // words parse into an f32 center; the radius/angle below then
+            // derive on the f32-rounded offsets, so borderline arcs' ceil
+            // segmentation must see the same rounded values.
+            let i = word(code, 'I').unwrap_or(0.0) as f32 as f64;
+            let j = word(code, 'J').unwrap_or(0.0) as f32 as f64;
             (i, j, (i * i + j * j).sqrt())
         }
     };
